Move layouts folder to src/lib/geometry
