+++ src/lib/3d/KeyboardKeyInstance.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { hoveredKey, clickedKey, clickedVisualSide, selectMode, protoConfig } from '$lib/store'
+  import { hoveredKey, clickedKey, clickedSide, selectMode, protoConfig } from '$lib/store'
   import { getContext } from 'svelte'
   import { nthKey, type CosmosKeyboard } from '$lib/worker/config.cosmos'
   import { Instance, interactivity } from '@threlte/extras'
@@ -10,11 +10,6 @@
   export let position: Vector3Tuple
   export let rotation: Vector4Tuple
   export let side: 'left' | 'right' | 'unibody' = 'right'
-
-  function onClick() {
-    $clickedKey = index
-    $clickedVisualSide = side
-  }
 
   type InteractivityContext = ReturnType<typeof interactivity>
   const context: InteractivityContext = getContext('interactivity')
@@ -47,11 +42,14 @@
   quaternion={rotation}
   on:pointerenter={() => ($hoveredKey = index)}
   on:pointerleave={() => ($hoveredKey = null)}
-  on:click={onClick}
+  on:click={() => {
+    $clickedKey = index
+    $clickedSide = side
+  }}
   on:pointermissed={() => {
     if (!context.initialHits.length) {
       $clickedKey = null
-      $clickedVisualSide = null
+      $clickedSide = null
     }
   }}
 />
