+++ src/lib/worker/config.cosmos.ts
@@ -1,8 +1,7 @@
 import ETrsf, { Constant, fullMirrorETrsf, type MatrixOptions, mirror } from '$lib/worker/modeling/transformation-ext'
 // import { deserialize } from 'src/routes/beta/lib/serialize'
-import { flippedKey } from '$lib/geometry/keycaps'
+import { DEFAULT_LAYOUT, flipLetter, flippedKey, getLayout, isAlphaLetter, isNamedLayoutId, LAYOUT, type LayoutId, NAMED_LAYOUT_IDS, type NamedLayoutId } from '$lib/geometry/layouts'
 import { decodeVariant, encodeVariant, PART_INFO, socketSize } from '$lib/geometry/socketsParts'
-import { DEFAULT_LAYOUT, flipLetter, getLayout, isAlphaLetter, isNamedLayoutId, LAYOUT, type LayoutId, NAMED_LAYOUT_IDS, type NamedLayoutId } from '$lib/layouts'
 import { type ClusterName, type ClusterSide, type ClusterType, type Connector, decodeClusterFlags, encodeClusterFlags, type PartVariant, type ScrewFlags } from '$target/cosmosStructs'
 import type { Curvature } from '$target/proto/cosmos'
 import { Matrix4, Vector3 } from 'three'
