+++ src/lib/worker/config.serialize.ts
@@ -2,7 +2,7 @@
  * For now, the new serializationf format for Cosmos is kept in a separate file.
  */
 
-import { DEFAULT_LAYOUT, LAYOUT_ENCODE, type LayoutId } from '$lib/layouts'
+import { DEFAULT_LAYOUT, LAYOUT_ENCODE, type LayoutId } from '$lib/geometry/layouts'
 import { BinaryReader, BinaryWriter } from '@protobuf-ts/runtime'
 import {
   decodeBasicShellFlags,
